style: RuleMapを移設
